move method to util and update classes accordingly
