+++ metric_learn/base_metric.py
@@ -1,4 +1,3 @@
-from numpy.linalg import cholesky
 from sklearn.base import BaseEstimator
 from sklearn.utils.validation import _is_arraylike
 from sklearn.metrics import roc_auc_score
@@ -180,32 +179,6 @@ class MahalanobisMixin(six.with_metaclass(ABCMeta, BaseMetricLearner,
 
   def metric(self):
     return self.transformer_.T.dot(self.transformer_)
-
-  def transformer_from_metric(self, metric):
-    """Computes the transformation matrix from the Mahalanobis matrix.
-
-    Since by definition the metric `M` is positive semi-definite (PSD), it
-    admits a Cholesky decomposition: L = cholesky(M).T. However, currently the
-    computation of the Cholesky decomposition used does not support
-    non-definite matrices. If the metric is not definite, this method will
-    return L = V.T w^( -1/2), with M = V*w*V.T being the eigenvector
-    decomposition of M with the eigenvalues in the diagonal matrix w and the
-    columns of V being the eigenvectors. If M is diagonal, this method will
-    just return its elementwise square root (since the diagonalization of
-    the matrix is itself).
-
-    Returns
-    -------
-    L : (d x d) matrix
-    """
-
-    if np.allclose(metric, np.diag(np.diag(metric))):
-      return np.sqrt(metric)
-    elif not np.isclose(np.linalg.det(metric), 0):
-      return cholesky(metric).T
-    else:
-      w, V = np.linalg.eigh(metric)
-      return V.T * np.sqrt(np.maximum(0, w[:, None]))
 
 
 class _PairsClassifierMixin(BaseMetricLearner):
